+++ lib/src/easy_table_cell.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/widgets.dart';
 
+/// Default cell. Used by automatic cell builder.
 class EasyTableCell extends StatelessWidget {
   const EasyTableCell({Key? key, this.value, this.textStyle}) : super(key: key);
 
+  /// Builds a cell that maps the value to a [double].
   factory EasyTableCell.double(
       {Key? key, required double value, int? fractionDigits}) {
     String? str;
@@ -14,6 +16,7 @@ class EasyTableCell extends StatelessWidget {
     return EasyTableCell(key: key, value: str);
   }
 
+  /// Builds a cell that maps the value to a [int].
   factory EasyTableCell.int({Key? key, required int value}) {
     return EasyTableCell(key: key, value: value.toString());
   }