+++ lib/src/easy_table_cell_builder.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 
-/// Signature for a function that creates a widget for a given column and row.
+/// Signature for a function that builds a widget for a given column and row.
 ///
-/// Used by [EasyTable].
+/// Used by [EasyTableColumn].
 typedef EasyTableCellBuilder<ROW> = Widget Function(
     BuildContext context, ROW row);