+++ lib/src/easy_table_row_color.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 
+/// Signature for a function that defines a row color.
 typedef EasyTableRowColor<ROW> = Color? Function(ROW row, int rowIndex);
 
+/// Default row colors rules.
 class RowColors {
   static EasyTableRowColor evenOdd({Color? evenColor, Color? oddColor}) {
     return (row, rowIndex) {