+++ lib/src/easy_table_value_mapper.dart
@@ -1,1 +1,4 @@
+/// Signature for a function that maps the value of a row.
+///
+/// Used by [EasyTableColumn].
 typedef EasyTableValueMapper<ROW> = dynamic Function(ROW row);