+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.0
+
+* Initial release
+  * Bidirectional scroll bars
+  * Columns
+    * Initial width
+    * Header builder
+    * Cell builder
+      * Initial automatic cell builder with data mapper
+
 ## 0.0.1
 
 * Package creation